# docs/swagger.yaml
basePath: /1sat
definitions:
...
  github_com_b-open-io_1sat-stack_pkg_txo.IndexedOutput:
    type: object
...
  pkg_bsv21.BalanceResponse:
    properties:
      balance:
        type: integer
      utxoCount:
        type: integer
    type: object
  pkg_bsv21.BlockInfo:
    properties:
      hash:
        type: string
      height:
        type: integer
      previousblockhash:
        type: string
      timestamp:
        type: integer
    type: object
  pkg_bsv21.BlockResponse:
    properties:
      block:
        $ref: '#/definitions/pkg_bsv21.BlockInfo'
      transactions:
        items:
          $ref: '#/definitions/pkg_bsv21.TransactionData'
        type: array
    type: object
  pkg_bsv21.TokenResponse:
    properties:
      amt:
        type: string
      dec:
        type: integer
      icon:
        type: string
      id:
        type: string
      op:
        type: string
      sym:
        type: string
      txid:
        type: string
      vout:
        type: integer
    type: object
  pkg_bsv21.TransactionData:
    properties:
      beef:
        items:
          type: integer
        type: array
      block_height:
        type: integer
      outputs:
        items:
          $ref: '#/definitions/github_com_b-open-io_1sat-stack_pkg_txo.IndexedOutput'
        type: array
      txid:
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/b-open-io/1sat-stack
  description: Composable BSV blockchain services API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 1Sat Stack API
  version: "1.0"
paths:
...
  /bsv21/{tokenId}:
    get:
      parameters:
      - description: 'Token ID (outpoint format: txid_vout)'
        in: path
        name: tokenId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_bsv21.TokenResponse'
      summary: Get BSV21 token details
      tags:
      - bsv21
  /bsv21/{tokenId}/{lockType}/{address}/balance:
    get:
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Lock type (p2pkh, cos, list, etc.)
        in: path
        name: lockType
        required: true
        type: string
      - description: Address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_bsv21.BalanceResponse'
      summary: Get address token balance
      tags:
      - bsv21
  /bsv21/{tokenId}/{lockType}/{address}/history:
    get:
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Lock type
        in: path
        name: lockType
        required: true
        type: string
      - description: Address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_b-open-io_1sat-stack_pkg_txo.IndexedOutput'
            type: array
      summary: Get address transaction history
      tags:
      - bsv21
  /bsv21/{tokenId}/{lockType}/{address}/unspent:
    get:
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Lock type
        in: path
        name: lockType
        required: true
        type: string
      - description: Address
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_b-open-io_1sat-stack_pkg_txo.IndexedOutput'
            type: array
      summary: Get address unspent outputs
      tags:
      - bsv21
  /bsv21/{tokenId}/{lockType}/balance:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Lock type
        in: path
        name: lockType
        required: true
        type: string
      - description: Array of addresses (max 100)
        in: body
        name: addresses
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_bsv21.BalanceResponse'
      summary: Get multi-address token balance
      tags:
      - bsv21
  /bsv21/{tokenId}/{lockType}/history:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Lock type
        in: path
        name: lockType
        required: true
        type: string
      - description: Array of addresses (max 100)
        in: body
        name: addresses
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_b-open-io_1sat-stack_pkg_txo.IndexedOutput'
            type: array
      summary: Get multi-address transaction history
      tags:
      - bsv21
  /bsv21/{tokenId}/{lockType}/unspent:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Lock type
        in: path
        name: lockType
        required: true
        type: string
      - description: Array of addresses (max 100)
        in: body
        name: addresses
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_b-open-io_1sat-stack_pkg_txo.IndexedOutput'
            type: array
      summary: Get multi-address unspent outputs
      tags:
      - bsv21
  /bsv21/{tokenId}/blk/{height}:
    get:
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Block height
        in: path
        name: height
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pkg_bsv21.BlockResponse'
      summary: Get block data for a token
      tags:
      - bsv21
  /bsv21/{tokenId}/tx/{txid}:
    get:
      parameters:
      - description: Token ID
        in: path
        name: tokenId
        required: true
        type: string
      - description: Transaction ID
        in: path
        name: txid
        required: true
        type: string
      - description: Include BEEF data
        in: query
        name: beef
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get transaction details
      tags:
      - bsv21
...
swagger: "2.0"